src/requirements.txt
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
scikit-image==0.21.0
